# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Jimma City Digital Street Address MIS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /streets/{id}/maintenance:
    get:
      operationId: getStreetMaintenanceHistory
      tags: [road-inventory]
      summary: Get maintenance history and inspection logs for a specific street
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Maintenance history of the street
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoadMaintenanceRecord"
...
  /road-inventory/records:
    get:
      operationId: listRoadMaintenanceRecords
      tags: [road-inventory]
      summary: List road maintenance and inspection records
      parameters:
        - name: streetId
          in: query
          schema:
            type: integer
        - name: activityType
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of maintenance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoadMaintenanceRecord"
    post:
      operationId: createRoadMaintenanceRecord
      tags: [road-inventory]
      summary: Log a new maintenance activity or PCI inspection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoadMaintenanceRecordInput"
      responses:
        "201":
          description: Created maintenance record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoadMaintenanceRecord"
  /road-inventory/records/{id}:
    put:
      operationId: updateRoadMaintenanceRecord
      tags: [road-inventory]
      summary: Update a maintenance record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoadMaintenanceRecordInput"
      responses:
        "200":
          description: Updated maintenance record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoadMaintenanceRecord"
        "404":
          description: Record not found
    delete:
      operationId: deleteRoadMaintenanceRecord
      tags: [road-inventory]
      summary: Delete a maintenance record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted successfully
        "404":
          description: Record not found
...
components:
  schemas:
...
    RoadMaintenanceRecord:
      type: object
      required: [id, streetId, activityType, performedDate, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        streetId:
          type: integer
        streetName:
          type: ["string", "null"]
        streetCode:
          type: ["string", "null"]
        kebeleName:
          type: ["string", "null"]
        activityType:
          type: string
          enum: [inspection, resurfacing, pothole_patching, drainage_clearing, lighting_repair, expansion, emergency_repair]
        pciScore:
          type: ["integer", "null"]
        pciRating:
          type: ["string", "null"]
          enum: [good, satisfactory, fair, poor, very_poor, serious, failed]
        distressTypes:
          type: ["string", "null"]
        performedDate:
          type: string
        contractor:
          type: ["string", "null"]
        costEtb:
          type: ["number", "null"]
        fundingSource:
          type: ["string", "null"]
          enum: [municipal_budget, regional_grant, federal_grant, community_fund]
        nextInspectionDue:
          type: ["string", "null"]
        status:
          type: string
          enum: [completed, in_progress, scheduled, deferred]
        inspectorName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    RoadMaintenanceRecordInput:
      type: object
      required: [streetId, activityType, performedDate]
      properties:
        streetId:
          type: integer
        activityType:
          type: string
          enum: [inspection, resurfacing, pothole_patching, drainage_clearing, lighting_repair, expansion, emergency_repair]
        pciScore:
          type: ["integer", "null"]
        pciRating:
          type: ["string", "null"]
          enum: [good, satisfactory, fair, poor, very_poor, serious, failed]
        distressTypes:
          type: ["string", "null"]
        performedDate:
          type: string
        contractor:
          type: ["string", "null"]
        costEtb:
          type: ["number", "null"]
        fundingSource:
          type: ["string", "null"]
          enum: [municipal_budget, regional_grant, federal_grant, community_fund]
        nextInspectionDue:
          type: ["string", "null"]
        status:
          type: string
          enum: [completed, in_progress, scheduled, deferred]
        inspectorName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]