# lib/api-zod/src/generated/types/roadMaintenanceRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Jimma City Digital Street Address MIS API
 * OpenAPI spec version: 0.1.0
 */
import type { RoadMaintenanceRecordActivityType } from './roadMaintenanceRecordActivityType';
import type { RoadMaintenanceRecordFundingSource } from './roadMaintenanceRecordFundingSource';
import type { RoadMaintenanceRecordPciRating } from './roadMaintenanceRecordPciRating';
import type { RoadMaintenanceRecordStatus } from './roadMaintenanceRecordStatus';

export interface RoadMaintenanceRecord {
  id: number;
  streetId: number;
  /** @nullable */
  streetName?: string | null;
  /** @nullable */
  streetCode?: string | null;
  /** @nullable */
  kebeleName?: string | null;
  activityType: RoadMaintenanceRecordActivityType;
  /** @nullable */
  pciScore?: number | null;
  /** @nullable */
  pciRating?: RoadMaintenanceRecordPciRating;
  /** @nullable */
  distressTypes?: string | null;
  performedDate: string;
  /** @nullable */
  contractor?: string | null;
  /** @nullable */
  costEtb?: number | null;
  /** @nullable */
  fundingSource?: RoadMaintenanceRecordFundingSource;
  /** @nullable */
  nextInspectionDue?: string | null;
  status: RoadMaintenanceRecordStatus;
  /** @nullable */
  inspectorName?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}