# lib/api-zod/src/generated/types/roadMaintenanceRecordInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Jimma City Digital Street Address MIS API
 * OpenAPI spec version: 0.1.0
 */
import type { RoadMaintenanceRecordInputActivityType } from './roadMaintenanceRecordInputActivityType';
import type { RoadMaintenanceRecordInputFundingSource } from './roadMaintenanceRecordInputFundingSource';
import type { RoadMaintenanceRecordInputPciRating } from './roadMaintenanceRecordInputPciRating';
import type { RoadMaintenanceRecordInputStatus } from './roadMaintenanceRecordInputStatus';

export interface RoadMaintenanceRecordInput {
  streetId: number;
  activityType: RoadMaintenanceRecordInputActivityType;
  /** @nullable */
  pciScore?: number | null;
  /** @nullable */
  pciRating?: RoadMaintenanceRecordInputPciRating;
  /** @nullable */
  distressTypes?: string | null;
  performedDate: string;
  /** @nullable */
  contractor?: string | null;
  /** @nullable */
  costEtb?: number | null;
  /** @nullable */
  fundingSource?: RoadMaintenanceRecordInputFundingSource;
  /** @nullable */
  nextInspectionDue?: string | null;
  status?: RoadMaintenanceRecordInputStatus;
  /** @nullable */
  inspectorName?: string | null;
  /** @nullable */
  notes?: string | null;
}